Код создания карты по изображению распределен на несколько классов
