+++ src/RevitDocumenter/Models/RevitRepository.cs
@@ -92,4 +92,25 @@ internal class RevitRepository {
         }
         return refs;
     }
+
+    public int GetMinDimensionInView() {
+        int scale = Document.ActiveView.Scale;
+        // При масштабе 1:100 корректно смотрится на виде размер до 2 футов
+        int minDimension = 2;
+        // Ищем минимальное значение размера, которое будет корректно видимо на виде с учетом масштаба
+        //      100       scale         minDimension * scale
+        // ------------ = -----  => x = --------------------
+        // minDimension     x                   100
+        return minDimension * scale / 100;
+    }
+
+    public void DeleteElementsById(List<ElementId> list) {
+        using var subTransaction = new SubTransaction(Document);
+        subTransaction.Start();
+
+        foreach(var id in list) {
+            Document.Delete(id);
+        }
+        subTransaction.Commit();
+    }
 }
